backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
openai==1.10.0
google-generativeai==0.3.2
langchain==0.1.4
langgraph==0.0.20
langchain-openai==0.0.5
langchain-google-genai==0.0.6

# Database - Updated for compatibility
supabase==2.7.4
httpx==0.27.0

# WebSocket
websockets==12.0
python-multipart==0.0.6

# Utilities
python-jose[cryptography]==3.3.0
